+++ packages/data-model/index.ts
@@ -1,1 +1,2 @@
+export * from './src/functions/migrate.js';
 export * from './src/models/index.js';
